instantiate type combinations

+++ cpp/src/community/k_truss_mg_v64_e64.cu
@@ -20,22 +20,22 @@ namespace cugraph {
 
 // MG instantiation
 
-template std::tuple<rmm::device_uvector<int32_t>,
-                    rmm::device_uvector<int32_t>,
+template std::tuple<rmm::device_uvector<int64_t>,
+                    rmm::device_uvector<int64_t>,
                     std::optional<rmm::device_uvector<float>>>
 k_truss(raft::handle_t const& handle,
-        graph_view_t<int32_t, int32_t, false, true> const& graph_view,
-        std::optional<edge_property_view_t<int32_t, float const*>> edge_weight_view,
-        int32_t k,
+        graph_view_t<int64_t, int64_t, false, true> const& graph_view,
+        std::optional<edge_property_view_t<int64_t, float const*>> edge_weight_view,
+        int64_t k,
         bool do_expensive_check);
 
-template std::tuple<rmm::device_uvector<int32_t>,
-                    rmm::device_uvector<int32_t>,
+template std::tuple<rmm::device_uvector<int64_t>,
+                    rmm::device_uvector<int64_t>,
                     std::optional<rmm::device_uvector<double>>>
 k_truss(raft::handle_t const& handle,
-        graph_view_t<int32_t, int32_t, false, true> const& graph_view,
-        std::optional<edge_property_view_t<int32_t, double const*>> edge_weight_view,
-        int32_t k,
+        graph_view_t<int64_t, int64_t, false, true> const& graph_view,
+        std::optional<edge_property_view_t<int64_t, double const*>> edge_weight_view,
+        int64_t k,
         bool do_expensive_check);
 
 }  // namespace cugraph
